Split executable and ui lib
