+++ crates/lavagna-bevy/.cargo/aarch64-linux-android.config.toml
@@ -1,0 +1,2 @@
+[build]
+target = "aarch64-linux-android"
