+++ crates/lavagna-bevy/.gitignore
@@ -1,0 +1,8 @@
+/target
+
+
+# Added by cargo
+#
+# already existing elements were commented out
+
+#/target
