+++ crates/lavagna/Cargo.toml
@@ -3,16 +3,6 @@ name = "lavagna"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-bevy = { version = "0.10", features = ["dynamic_linking"] }
-
-# Enable a small amount of optimization in debug mode
-[profile.dev]
-opt-level = 1
-
-# Enable high optimizations for dependencies (incl. Bevy), but not for our code:
-[profile.dev.package."*"]
-opt-level = 3
+lavagna-bevy = { path = "../lavagna-bevy" }
